Finalize ProjectInformationLayout and wire ViewModel Information to ProjectInformationLayout
